source-mysql: use datatype integration test fixture
